[Projec4-Bugs] (Add) Create services

+++ controllers/chapters.php
@@ -2,6 +2,7 @@
 
 include_once 'models/chapters.php';
 include_once 'models/comments.php';
+include_once 'services/auth.php';
 
 function chapters($page) {
     $chapter = new Chapters();
@@ -12,7 +13,7 @@ function chapters($page) {
 
 function showChapter($page) {
     require 'views/includes/bootstrap.php';
-    if(isset($_SESSION['auth'])) {
+    if(is_admin()) {
         $chapter = new Chapters();
         $showChapter = $chapter->showChapter();
         $date = $showChapter['created_at'];
@@ -21,44 +22,52 @@ function showChapter($page) {
         $comment = new Comments();
         $allComments = $comment->getAllComments($showChapter['id']);
         include_once 'views/chapters/sections/chapter/chapter-show/'.$page.'_show_view.php';
-    } else {
-        header('Location: /projet4/login?action=login'); 
-        Session::getInstance()->setFlash('info', "Pour visualiser le chapitre en entier veuillez vous connecter.");         
     }
-  
 }
 
 function addChapter($page) { 
     require 'views/includes/bootstrap.php';
-    if(!empty($_POST) && isset($_POST['btnAjoutTicket'])) {  
-        if(!empty($_POST['titleTicket']) && !empty($_POST['contentTicket'])) {
-            $titleTicket = str_secur($_POST['titleTicket']);
-            $contentTicket = str_secur($_POST['contentTicket']);
-            $chapter = new Chapters();
-            $addChapter = $chapter->addChapter($_POST['titleTicket'], $_POST['contentTicket'], $_SESSION['auth']->id);
-            Session::getInstance()->setFlash('success', "Le chapitre a bien été ajouté.");
-            header('Location: /projet4/chapters?action=chapters'); exit;
-        }
+    if(is_admin()) {
+        if(is_role_admin()) {
+            if(!empty($_POST) && isset($_POST['btnAjoutTicket'])) {  
+                if(!empty($_POST['titleTicket']) && !empty($_POST['contentTicket'])) {
+                    $titleTicket = str_secur($_POST['titleTicket']);
+                    $contentTicket = str_secur($_POST['contentTicket']);
+                    $chapter = new Chapters();
+                    $addChapter = $chapter->addChapter($_POST['titleTicket'], $_POST['contentTicket'], $_SESSION['auth']->id);
+                    Session::getInstance()->setFlash('success', "Le chapitre a bien été ajouté.");
+                    header('Location: /projet4/chapters?action=chapters'); exit;
+                }
+            }
+            include_once 'views/chapters/sections/chapter/chapter-add/'.$page.'_add_view.php';     
+        } 
     }
-    include_once 'views/chapters/sections/chapter/chapter-add/'.$page.'_add_view.php';   
 }
 
 function deleteChapter($page) { 
     require 'views/includes/bootstrap.php';
-    $chapter = new Chapters();
-    $deleteChapter = $chapter->deleteChapter();
-    Session::getInstance()->setFlash('success', "Le chapitre a bien été supprimé.");
-    header('Location: /projet4/chapters?action=chapters');  
+    if(is_admin()) {
+        if(is_role_admin()) {
+            $chapter = new Chapters();
+            $deleteChapter = $chapter->deleteChapter();
+            Session::getInstance()->setFlash('success', "Le chapitre a bien été supprimé.");
+            header('Location: /projet4/chapters?action=chapters');  
+        }
+    }
 }
 
 function updateChapter($page) {
     require 'views/includes/bootstrap.php';
-    $chapter = new Chapters();
-    $showChapter = $chapter->showChapter();
-    if(!empty($_POST) && isset($_POST['btnUpdateChapter'])) {
-        $updateChapter = $chapter->updateChapter($_POST['titleTicket'], $_POST['contentTicket'], $_GET['id']);
-        Session::getInstance()->setFlash('success', "Le chapitre a bien été mis à jour.");
-        header('Location: /projet4/chapters?action=showChapter&id='.$_GET['id']); exit;
+    if(is_admin()) {
+        if(is_role_admin()) {
+            $chapter = new Chapters();
+            $showChapter = $chapter->showChapter();
+            if(!empty($_POST) && isset($_POST['btnUpdateChapter'])) {
+                $updateChapter = $chapter->updateChapter($_POST['titleTicket'], $_POST['contentTicket'], $_GET['id']);
+                Session::getInstance()->setFlash('success', "Le chapitre a bien été mis à jour.");
+                header('Location: /projet4/chapters?action=showChapter&id='.$_GET['id']); exit;
+            }
+            include_once 'views/chapters/sections/chapter/chapter-update/'.$page.'_update_view.php'; 
+        }
     }
-    include_once 'views/chapters/sections/chapter/chapter-update/'.$page.'_update_view.php'; 
 }
